requirements.txt
aiohttp==3.7.2
aiohttp-basicauth==0.2.0
aiohttp-jinja2==1.3.0
aiosmtpd==1.2
aiosqlite==0.16.0
alembic==1.0.0.dev0
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
arabic-reshaper==2.1.3
arrow==0.15.5
asgiref==3.2.5
asn1crypto==0.24.0
astroid==2.3.3
async-timeout==3.0.1
atpublic==2.1.1
attrs==20.3.0
autopep8==1.5.2
Babel==2.6.0
bcrypt==3.1.6
beautifulsoup4==4.9.1
billiard==3.3.0.23
binaryornot==0.4.4
bleach==3.1.5
blinker==1.4
Brlapi==0.6.7
celery==3.1.26.post2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.3.7
configobj==5.0.6
confusable-homoglyphs==3.2.0
cookiecutter==1.7.2
cryptography==2.8
cssmin==0.2.0
cssprefixer==1.4.1
cssutils==1.0.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.3.0
Django==2.2.20
django-admin-thumbnails==0.2.5
django-allauth==0.42.0
django-braces==1.13.0
django-ckeditor==6.0.0
django-colorfield==0.3.0
django-cors-headers==3.10.0
django-countries==6.1.2
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-extensions==2.2.8
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-formtools==2.2
django-google-translate==1.1
django-js-asset==1.2.2
django-localflavor==3.0.1
django-mptt==0.11.0
django-multiselectfield==0.1.12
django-registration==3.1
django-resized==0.3.11
django-rest-auth==0.9.5
django-tinymce==3.0.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.16
easy-thumbnails==2.7
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-Assets==2.0
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.4.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
future==0.16.0
gevent==20.6.2
gevent-socketio-hartwork==0.3.6.post201803091952
gevent-websocket==0.10.1
greenlet==0.4.16
gunicorn==20.1.0
html5lib==1.0.1
httplib2==0.11.3
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==0.24
Jinja2==2.11.2
jinja2-time==0.2.0
keyring==17.1.1
keyrings.alt==3.1.1
kombu==3.0.37
lazy-object-proxy==1.4.3
ldap3==2.4.1
lockfile==0.12.2
Logbook==1.5.3
louis==3.8.0
mailtrap==0.1.6
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
multidict==5.0.0
oauthlib==3.1.0
olefile==0.46
packaging==20.9
paramiko==2.4.2
passlib==1.7.1
pexpect==4.6.0
Pillow==7.2.0
pipenv==2018.11.26
poyo==0.5.0
psutil==5.5.1
psycopg2==2.8.6
pyasn1==0.4.2
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.2
PyGObject==3.30.4
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.4.4
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyScss==1.3.7
PySimpleSOAP==1.16.2
python-apt==1.8.4.3
python-bidi==0.4.2
python-daemon==2.2.4
python-dateutil==2.8.1
python-debian==0.1.35
python-debianbts==2.8.2
python-editor==1.0.3
python-resize-image==1.1.19
python-slugify==4.0.0
python-stdnum==1.13
python3-openid==3.1.0
pytz==2019.3
pyxdg==0.25
PyYAML==3.13
redis==3.4.1
reportbug===7.5.3-deb10u1
reportlab==3.5.13
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.15.34
SecretStorage==2.3.1
sendria==2.0.1
simplejson==3.16.0
six==1.12.0
social-auth-core==3.3.3
sorl-thumbnail==12.5.0
soupsieve==2.0.1
speaklater==1.4
SQLAlchemy==1.2.18
sqlparse==0.3.1
sshtunnel==0.1.4
stripe==2.44.0
structlog==20.1.0
text-unidecode==1.3
tinify==1.5.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.8
virtualenv==20.0.17
virtualenv-clone==0.5.4
webassets==2.0
webencodings==0.5.1
Werkzeug==1.0.0
whichcraft==0.4.1
whitenoise==5.3.0
wrapt==1.11.2
WTForms==2.2.1
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
yarl==1.6.2
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
